# arcanium/api/openapi/arcanium.yaml
openapi: 3.1.0
info:
  title: Arcanium API
  version: "1.0.0"
  description: >
    Arcanium's cryptographic control-plane API. Authored from the actual
    Express route handlers (arcanium/api/src/routes/*.js,
    arcanium/api/src/auth/index.js) as source of truth for current
    behavior — a capture-then-enforce exercise (Prompt 22), not a redesign.
    This is the ONE authoritative contract: the Nuxt gateway allowlist
    (arcanium/ui/server/routes/gateway/[...path].ts), the CLI request
    shapes, and docs/api.md's narrative must all be checked against this
    file, not maintained as independent parallel sources of truth.
servers:
  - url: http://localhost:3001
    description: Direct API (dev/scenario scripts)
  - url: http://localhost:3000/gateway
    description: Same-origin UI gateway (browser-facing; allowlisted subset only)
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: arc_session
      description: >
        Opaque session id issued by GET /api/v1/auth/callback. When
        ARCANIUM_AUTH_ENABLED is unset, requireSession no-ops and every
        request is treated as an estate-wide operator — see
        docs/security.md.
    serviceAccountBearer:
      type: http
      scheme: bearer
      description: >
        Reserved for a future machine-to-machine identity source
        (prompts/26_external_integration_key_lifecycle.md, not yet
        executed). Not accepted by any route in this version of the API —
        listed here only so the contract does not need a breaking change
        once it is.
  responses:
    Unauthorized:
      description: No session cookie, or the session is expired/invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: >
        Authenticated, but authorize() denied the action — wrong role,
        wrong tenant, or a scoped grant that did not match (Phase 18's
        deny-by-default matrix, auth/authorize.js).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: >
        Missing, or deliberately indistinguishable from missing for a
        cross-tenant read (Phase 18's GET /applications/:id precedent —
        a supplier-admin gets 404, not 403, for another tenant's resource,
        so existence itself isn't leaked).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Request failed field validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error:
          description: Direct-API responses use a string; gateway responses use boolean true.
          oneOf:
            - type: string
            - type: boolean
        statusCode: { type: integer }
        statusMessage: { type: string }
        message: { type: string }
        field: { type: string, description: "Direct-API 400 responses name the offending field." }
        reason: { type: string, description: "authorize() denial reason, on 403 responses." }
        action: { type: string, description: "The authorize() action that was denied, on 403 responses." }
      required: [error]

    # ── Suppliers / Applications ─────────────────────────────────────────
...
    CryptoProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        application_id: { type: string, format: uuid }
        type: { type: string, enum: [transit, pki, kmip, managed_key] }
        vault_path: { type: string }
        custody: { type: string }
        rotation_days: { type: ["integer", "null"] }
        created_at: { type: string, format: date-time }
      required: [id, application_id, type, vault_path, created_at]
    Application:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: ["string", "null"] }
        supplier_id: { type: ["string", "null"], format: uuid }
        category: { type: string, enum: [platform, tenant, unscoped] }
        environment:
          type: string
          description: "Prompt 27 — defaults to 'production' when omitted at registration; explicitly recorded, never left null."
        registered_at: { type: string, format: date-time }
        crypto_profiles:
          type: array
          items: { $ref: "#/components/schemas/CryptoProfile" }
      required: [id, name, registered_at]

    # ── Teams (Prompt 27, control-plane multi-tenancy) ────────────────────
...
    ProvisioningJob:
      type: object
      properties:
        id: { type: string, format: uuid }
        target_type: { type: string, enum: [supplier, application, key] }
        target_id: { type: string }
        target_name: { type: ["string", "null"] }
        action: { type: string }
        status:
          type: string
          enum: [pending, running, succeeded, rolling_back, rolled_back, rollback_failed, failed]
        requested_by: { type: string }
        params: { type: object, additionalProperties: true }
        steps: { type: array, items: { type: object, additionalProperties: true } }
        error: { type: ["string", "null"] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, status]

    # ── Approvals ─────────────────────────────────────────────────────────
...
    ObservationStatus:
      type: string
      enum: [COMPLIANT, DRIFTED, UNKNOWN]
    Disposition:
      type: string
      enum: [OPEN, EXCEPTION_ACCEPTED, RECONCILED]
...
    ControlStatus:
      type: string
      enum: [PASS, FAIL, UNKNOWN, "N/A"]
    ConfidenceLevel:
      type: string
      enum: [HIGH, MEDIUM, LOW]
    ControlAssessment:
      type: object
      properties:
        id: { type: string, format: uuid }
        control_id: { type: string, example: "ROT-POL-01" }
        scope: { type: string, example: "suppliers/pepsi/payments-api/payments-api-key" }
        status: { $ref: "#/components/schemas/ControlStatus" }
        desired_value: { type: ["object", "null"], additionalProperties: true }
        observed_value: { type: ["object", "null"], additionalProperties: true }
        evidence_refs: { type: object, additionalProperties: true }
        freshness_seconds: { type: ["integer", "null"] }
        confidence: { $ref: "#/components/schemas/ConfidenceLevel" }
        assessed_at: { type: string, format: date-time }
        requirement: { type: string }
        mandatory: { type: boolean }
        dimension: { type: string }
      required: [id, control_id, scope, status, confidence, assessed_at]
...
    IntentCustodyEntry:
      type: object
      properties:
        profile_id: { type: string, format: uuid }
        key_name: { type: string }
        vault_path: { type: string }
        type: { type: string, enum: [transit, pki, kmip, managed_key] }
        custody: { type: string }
        hsm_backed: { type: ["boolean", "null"], description: "null = not live-confirmable right now, never guessed" }
        rotation_days: { type: ["integer", "null"] }
      required: [profile_id, key_name, vault_path, type, custody]
    IntentReconciliationEntry:
      type: object
      properties:
        desired_state_id: { type: string, format: uuid }
        key_name: { type: string }
        requirement: { type: string }
        observation_status: { $ref: "#/components/schemas/ObservationStatus" }
        disposition: { $ref: "#/components/schemas/Disposition" }
        observed_at: { type: ["string", "null"], format: date-time }
      required: [desired_state_id, key_name, requirement, observation_status, disposition]
    IntentView:
      type: object
      properties:
        application_id: { type: string, format: uuid }
        application: { type: string }
        tenant: { type: string, example: "root" }
        environment:
          type: ["string", "null"]
          description: "Always null — Arcanium has no per-application environment concept in the data model yet. Never fabricated as a default like \"production\"."
        requirements:
          type: object
          properties:
            encryption: { type: boolean }
            signing: { type: boolean }
            kmip: { type: boolean }
            tls: { type: boolean }
          required: [encryption, signing, kmip, tls]
        custody:
          type: array
          items: { $ref: "#/components/schemas/IntentCustodyEntry" }
        governance:
          type: object
          properties:
            rotation_policies:
              type: array
              items:
                type: object
                properties:
                  desired_state_id: { type: string, format: uuid }
                  key_name: { type: string }
                  desired_days: { type: ["integer", "null"] }
                  source: { type: string }
                  changed_by: { type: string }
                  changed_reason: { type: ["string", "null"] }
                  updated_at: { type: string, format: date-time }
            destruction_requires_approval:
              type: boolean
              description: "A code invariant (provisioner/key.js's requestKeyDestroy() always creates an approval), not per-application configurable data — true for every application today."
            approvals:
              type: array
              items: { type: object, additionalProperties: true }
          required: [rotation_policies, destruction_requires_approval, approvals]
        desired_state:
          type: array
          items: { type: object, additionalProperties: true, description: "Raw desired_state rows, current version only" }
        observed_state:
          type: array
          items: { type: object, additionalProperties: true, description: "Latest reconciliation_runs row per desired_state row" }
        assessment:
          type: object
          properties:
            reconciliation:
              type: array
              items: { $ref: "#/components/schemas/IntentReconciliationEntry" }
            controls:
              type: array
              items: { $ref: "#/components/schemas/ControlAssessment" }
          required: [reconciliation, controls]
        evidence:
          type: array
          items:
            type: object
            properties:
              kind: { type: string, enum: [lifecycle_event, evidence] }
            additionalProperties: true
        entry_story:
          type: object
          description: "Prompt 25 Deliverable 3 — computed from real data presence, never a static string."
          properties:
            labels:
              type: array
              items: { type: string, enum: ["Lifecycle view", "Governance view", "Tenant isolation view"] }
            summary: { type: string }
          required: [labels, summary]
      required:
        [
          application_id, application, tenant, environment, requirements,
          custody, governance, desired_state, observed_state, assessment,
          evidence, entry_story,
        ]
...
paths:
...
  /api/v1/applications:
    get:
      summary: List applications (tenant-scoped)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Application" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Register an application
      description: "authorize({action:'provision', tenant}) — Prompt 19 closed the missing-tenant-param gap that previously always denied supplier-admins here."
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                supplier_id: { type: ["string", "null"], format: uuid }
              required: [name]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Name already registered }
  /api/v1/applications/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Application detail (tenant-scoped since Phase 18 — the GET /applications/:id gap)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: Update registry metadata (Prompt 19 — previously completely unguarded)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { description: { type: ["string", "null"] } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Deprovision the application (Prompt 19 — previously completely unguarded)
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/applications/{id}/provision:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Provision the AppRole + policy + Transit key; seeds Phase 20 desired_state when rotation_days is set
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                custody: { type: string, enum: [vault, softhsm] }
                key_type: { type: string }
                capabilities: { type: array, items: { type: string } }
                rotation_days: { type: ["integer", "null"] }
                tls: { type: boolean }
      responses:
        "202":
          description: Accepted (queued or synchronously finished)
          content:
            application/json:
              schema:
                type: object
                properties:
                  application: { $ref: "#/components/schemas/Application" }
                  provisioning_job: { $ref: "#/components/schemas/ProvisioningJob" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/applications/{id}/classify:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      summary: Set platform/tenant/unscoped classification
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { category: { type: string, enum: [platform, tenant, unscoped] } }
              required: [category]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Application" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/applications/{id}/intent:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: Unified Cryptographic Service Intent View (Phase 25) — pure aggregation over Phases 18-21's existing data, tenant-scoped the same way GET /applications/{id} is
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IntentView" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── Keys ──────────────────────────────────────────────────────────────
...
  /api/v1/suppliers/{id}/applications:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      summary: This supplier's applications
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Application" } }
        "404": { $ref: "#/components/responses/NotFound" }
...